refactor: infrastructure 계층 도메인별로 폴더 분리
