+++ src/services/counselings/domains/counselors/counselors.module.ts
@@ -3,10 +3,10 @@ import { PersonaEntity } from "~shared/core/infrastructure/entities/counselors/P
 import { CounselorsPersister } from "~counselings/domains/counselors/counselors.persister";
 import { CounselorsReader } from "~counselings/domains/counselors/counselors.reader";
 import { CounselorsService } from "~counselings/domains/counselors/counselors.service";
-import { CounselorsRepository } from "~counselings/infrastructures/counselors.repository";
-import { PsqlCounselorsRepository } from "~counselings/infrastructures/psql-counselors.repository";
-import { RepositoryCounselorsPersister } from "~counselings/infrastructures/repository-counselors.persister";
-import { RepositoryCounselorsReader } from "~counselings/infrastructures/repository-counselors.reader";
+import { CounselorsRepository } from "~counselings/infrastructures/counselors/counselors.repository";
+import { PsqlCounselorsRepository } from "~counselings/infrastructures/counselors/psql-counselors.repository";
+import { RepositoryCounselorsPersister } from "~counselings/infrastructures/counselors/repository-counselors.persister";
+import { RepositoryCounselorsReader } from "~counselings/infrastructures/counselors/repository-counselors.reader";
 
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
